+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Each entry in this document is intended to be concise and focused on what is relevant to end-users.
 
+## [2.0.2] - 2024-02-17
+- Added support for multiple topics in the `listenMultiple` method.
+- Added support for pausing all subscriptions to multiple topics in the `pauseListeningToMultiple` method.
+
 ## [2.0.1] - 2023-12-19
 
 ### Added